Update 2024.9.22
- Fade는 GameObject에서 관리
